We added a new motor
Currently struggling on how to read the built in encoder of the motor.

Our goal was to be able to run two motors with the same button but one of them be staggered. So the fly wheel got up to speed before getting fed by the cone.

Will remove example subsystem stuff later. Simply wasn't imperative today